Moved `ak-sm-files` to module `smfiles`
Also, cleaning up the `bin/` directory:
- ak-sm-filejoiner and
- ak-sm-filesplitter
are now part of `lib/_ak_smfiles`

Updated README as well removing the above scripts and fixing the typo:
- sm-files -> smfiles
from the modules list.

Note that the module part, as well as the `files` module are going to be
removed in favor of `ak-fs`. Stay tuned for that.

To update, do:
```
git pull
./update.sh
```